# apps/coupon-api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDADAN Coupon API",
    "description": "상점별 쿠폰·도장 발급 시스템 API. 모든 경로는 `/api/coupon/v1` 아래에 있으며, JSON 필드는 snake_case, 식별자는 UUID 문자열, 시각은 RFC 3339 UTC, 금액은 원 단위 정수입니다.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:7810",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/coupon/v1/campaigns/{campaign_id}/claims": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "선착순 쿠폰 받기 (§11.3, CAMPAIGN-004).",
        "description": "The rate limit runs *before* the transaction, not after: SEC-003 wants a bot slowed\ndown before it can take stock, and is explicit that a legitimate customer must never\nhave a coupon revoked after the fact.",
        "operationId": "claim_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "Campaign id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Claimed, or the coupon already held",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimedCoupon"
                }
              }
            }
          },
          "409": {
            "description": "CAMPAIGN_SOLD_OUT"
          },
          "422": {
            "description": "Not eligible, or the campaign is not issuing"
          },
          "429": {
            "description": "Too many attempts"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
...
    "/api/coupon/v1/owner/campaigns": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "My campaigns, newest first, cursor-paginated (§11.1).",
        "operationId": "list_campaigns",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "1–100, default 20",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "next_cursor from the previous page",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaigns",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Campaign"
                }
              }
            }
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      },
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Draft a campaign (CAMPAIGN-001).",
        "operationId": "create_campaign",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "400": {
            "description": "A benefit, quantity or schedule value is out of range"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "One campaign.",
        "operationId": "get_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "Campaign id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "404": {
            "description": "No such campaign in this store"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      },
      "patch": {
        "tags": [
          "campaigns"
        ],
        "summary": "Edit a draft, or the forward-looking fields of a published campaign (CAMPAIGN-008).",
        "operationId": "patch_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "Campaign id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "description": "Expected version",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "The edit would apply retroactively to issued coupons"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/campaigns/{campaign_id}/cancel": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Cancel, naming what happens to coupons already issued (CAMPAIGN-007).",
        "operationId": "cancel_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "Campaign id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelCampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Already ended or cancelled"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/campaigns/{campaign_id}/estimate": {
      "get": {
        "tags": [
          "campaigns"
        ],
        "summary": "대상 예상 인원과 최대 발급 비용 (CAMPAIGN-002), before the confirmation modal.",
        "operationId": "estimate_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "Campaign id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "What publishing would commit to",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishEstimate"
                }
              }
            }
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/campaigns/{campaign_id}/pause": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Stop new issuance, keeping what customers already hold (CAMPAIGN-006).",
        "operationId": "pause_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "Campaign id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PauseCampaignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Paused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Not a campaign that can be paused"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/campaigns/{campaign_id}/publish": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Publish, and register the issuing job (§11.4, CAMPAIGN-003).",
        "description": "[`RecentlyAuthenticated`] rather than [`StoreOwner`]: CAMPAIGN-002 classes publishing\nas 고위험 작업 requiring re-authentication, and this is the same treatment §9.3 gives\nwithdrawal and business-identity edits.",
        "operationId": "publish_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "Campaign id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishedCampaign"
                }
              }
            }
          },
          "403": {
            "description": "Re-authentication is required for this action"
          },
          "422": {
            "description": "A publication rule is not satisfied"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
    "/api/coupon/v1/owner/campaigns/{campaign_id}/resume": {
      "post": {
        "tags": [
          "campaigns"
        ],
        "summary": "Continue from the checkpoint (CAMPAIGN-006).",
        "operationId": "resume_campaign",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "description": "Campaign id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "UUID, required on every mutation",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resumed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            }
          },
          "422": {
            "description": "Not a paused campaign, or the issuing period has ended"
          }
        },
        "security": [
          {
            "firebase": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelCampaignRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "revoke_policy": {
            "$ref": "#/components/schemas/RevokePolicy",
            "description": "CAMPAIGN-007: 발급 후 취소는 회수 정책을 명시한다."
          }
        }
      },
...
      "ClaimedCoupon": {
        "type": "object",
        "description": "The answer to `POST /campaigns/:id/claims` (§11.3).",
        "required": [
          "coupon_id",
          "campaign_id",
          "store_id",
          "title",
          "usable_from",
          "expires_at",
          "already_claimed"
        ],
        "properties": {
          "already_claimed": {
            "type": "boolean",
            "description": "True when this call returned a coupon the customer already had. §11.3 requires a\nrepeat request to answer with the existing coupon id rather than an error."
          },
          "campaign_id": {
            "type": "string",
            "format": "uuid"
          },
          "coupon_id": {
            "type": "string",
            "format": "uuid"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "remaining_quantity": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "What is left after this claim, or `None` while the campaign is unlimited."
          },
          "store_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "usable_from": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateCampaignRequest": {
        "type": "object",
        "description": "The campaign draft form (CAMPAIGN-001), in the order the wizard collects it:\n혜택 → 사용 조건 → 대상 → 수량 → 일정 → 알림.",
        "required": [
          "name",
          "customer_description",
          "benefit",
          "issue_mode",
          "total_quantity",
          "issue_starts_at",
          "issue_ends_at"
        ],
        "properties": {
          "allowed_local_time_ranges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LocalTimeRange"
            }
          },
          "allowed_weekdays": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "audience_criteria": {
            "$ref": "#/components/schemas/AudienceCriteria"
          },
          "audience_type": {
            "$ref": "#/components/schemas/AudienceType"
          },
          "benefit": {
            "$ref": "#/components/schemas/Benefit"
          },
          "customer_description": {
            "type": "string"
          },
          "eligible_category_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "eligible_item_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "excluded_item_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "external_discount_stackable": {
            "type": "boolean"
          },
          "issue_ends_at": {
            "type": "string",
            "format": "date-time"
          },
          "issue_mode": {
            "$ref": "#/components/schemas/IssueMode"
          },
          "issue_starts_at": {
            "type": "string",
            "format": "date-time"
          },
          "minimum_order_amount": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "notification_channels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notification_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "per_business_day_quantity": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "per_user_quantity": {
            "type": "integer",
            "format": "int32"
          },
          "relative_validity_days": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "restore_quantity_on_revoke": {
            "type": "boolean",
            "description": "§8.4: fixed on the campaign, default false. Not editable afterwards."
          },
          "total_quantity": {
            "$ref": "#/components/schemas/TotalQuantity"
          },
          "usable_from": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "usable_until": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
      "PauseCampaignRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PublishEstimate": {
        "type": "object",
        "description": "What publishing would cost, shown before the confirmation modal (CAMPAIGN-002).",
        "required": [
          "audience_size",
          "maximum_issued_quantity",
          "estimated_notification_count"
        ],
        "properties": {
          "audience_size": {
            "type": "integer",
            "format": "int64"
          },
          "estimated_notification_count": {
            "type": "integer",
            "format": "int64"
          },
          "maximum_discount_cost": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "maximum_issued_quantity": {
            "type": "integer",
            "format": "int64",
            "description": "최대 발급 비용의 참고값 — the ceiling, not a prediction."
          }
        }
      },
...
      "PublishedCampaign": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Campaign"
          },
          {
            "type": "object",
            "required": [
              "estimate"
            ],
            "properties": {
              "estimate": {
                "$ref": "#/components/schemas/PublishEstimate"
              },
              "job_id": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "uuid",
                "description": "The issuing job, for `DIRECT` campaigns. `None` for 선착순."
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "firebase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token"
      }
    }
  },
...
}